# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organization/catalog/items": {
      "get": {
        "operationId": "CatalogController_list",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/lookup": {
      "get": {
        "operationId": "CatalogController_lookup",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/units": {
      "get": {
        "operationId": "CatalogController_itemUnits",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_upsertItemUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/units/{unitId}": {
      "delete": {
        "operationId": "CatalogController_removeItemUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/barcodes": {
      "get": {
        "operationId": "CatalogController_itemBarcodes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_addItemBarcode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/barcodes/{barcode}": {
      "delete": {
        "operationId": "CatalogController_removeItemBarcode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/alternative-codes": {
      "get": {
        "operationId": "CatalogController_itemAlternativeCodes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_upsertItemAlternativeCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}/alternative-codes/{code}": {
      "delete": {
        "operationId": "CatalogController_removeItemAlternativeCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/organization/catalog/items/{id}": {
      "patch": {
        "operationId": "CatalogController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/organization/catalog/categories/{id}": {
      "patch": {
        "operationId": "CatalogController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_removeCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/organization/catalog/units/{id}": {
      "patch": {
        "operationId": "CatalogController_updateUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_removeUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/organization/catalog/tax-groups/{id}": {
      "patch": {
        "operationId": "CatalogController_updateTaxGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP SaaS API",
    "description": "Cloud multi-tenant ERP. Errors are RFC 9457 application/problem+json with a stable `code` from @erp/contracts. Tenant context always comes from the access token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "branchScope": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Branch-Id"
      },
      "idempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}